+++ R/drop_columns.R
@@ -28,6 +28,7 @@ drop_columns <- function(source, drops,
                          ...,
                          strict = TRUE,
                          env = parent.frame()) {
+  force(env)
   UseMethod("drop_columns", source)
 }
 
@@ -36,6 +37,7 @@ drop_columns.relop <- function(source, drops,
                                ...,
                                strict = TRUE,
                                env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)),
                           "rquery::drop_columns.relop")
   if(length(drops)<=0) {
@@ -60,6 +62,7 @@ drop_columns.data.frame <- function(source, drops,
                                     ...,
                                     strict = TRUE,
                                     env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)),
                           "rquery::drop_columns.data.frame")
   if(length(drops)<=0) {