+++ R/natural_join.R
@@ -61,6 +61,7 @@ natural_join <- function(a, b,
                          by,
                          jointype = 'INNER',
                          env = parent.frame()) {
+  force(env)
   UseMethod("natural_join", a)
 }
 
@@ -70,6 +71,7 @@ natural_join.relop <- function(a, b,
                                by,
                                jointype = 'INNER',
                                env = parent.frame()) {
+  force(env)
   if(length(list(...))>0) {
     stop("rquery::natural_join unexpected arguments")
   }
@@ -101,6 +103,7 @@ natural_join.data.frame <- function(a, b,
                                     by,
                                     jointype = 'INNER',
                                     env = parent.frame()) {
+  force(env)
   if(length(list(...))>0) {
     stop("rquery::natural_join unexpected arguments")
   }