+++ R/select_columns.R
@@ -23,11 +23,13 @@
 #' @export
 #'
 select_columns <- function(source, columns, env = parent.frame()) {
+  force(env)
   UseMethod("select_columns", source)
 }
 
 #' @export
 select_columns.relop <- function(source, columns, env = parent.frame()) {
+  force(env)
   if(length(columns)<=0) {
     stop("rquery::select_columns must select at least 1 column")
   }
@@ -43,6 +45,7 @@ select_columns.relop <- function(source, columns, env = parent.frame()) {
 
 #' @export
 select_columns.data.frame <- function(source, columns, env = parent.frame()) {
+  force(env)
   if(length(columns)<=0) {
     stop("rquery::select_columns must select at least 1 column")
   }